Add sight words context pack
This commit puts pre-k and kindergarten sight words into a context pack (as a pre-k word pack and k word pack). There are some words that are in pre-k in one form and in k as another form. For now, I have not combined all the forms together because it seems important for the words to be visible in their sight-word form (e.g., "see" and "saw" would normally be in "see" with forms "see", "sees", "saw", "seeing", but I've split them differently for now).