Update Unity and reimport the project
This is a branch to test the impact of reimporting the project and updating the version of Unity. It seems like maybe the meta files will be updated when the version of Unity changes, so I would like to check to see if that is the case. It would be annoying if the meta files update every time a different person edits something. Apparently, the meta files need to be here, so I'm just trying to make them less annoying while still letting them do their job.