Fix wordtiles duplicating when dragged
Previously if you dragged a tile from the wordbank to itself or the "default" space, you would create new game objects or duplicate the word tile each time.

The original copy is still being dragged in place of the clone though, so the word bank will have the dragged word tile with (clone) appended to it each time you move it.