Change the TTS for the sentence bar to read each word individually
I've created a new method: "startSpeakingSentenceSlowly" that mimics the coroutine we use for highlighting individual tiles and applied it to a foreach loop that goes through all the word tiles in the sentence bar.

I've also commented out some old bits of code in the TextToSpeechHandler.cs file that seemed to be breaking things.