Create the rough backbone for animating sentence creation.
I've changed the ClearTiles method to function as a coroutine that animates the tiles before destroying them. Currently it just bumps the tile up 20 units, but you could stick an animation in there and it should work just fine.

Going forward we need an animation for the tiles being sucked into the pipe and an animation for the conveyor belt spinning. We also would want to make the output sentence appear one word at a time, rather than all at once.

#32