Implement smoother animations and match them to TTS
This seems to work pretty well, although the actual implementation of this feature has gotten really messy. We should probably look into LeanTwean for making the animation scripts easier to use.

A known issue with this feature as it stands is that it doesn't move the scrollbar when you toggle the lever to submit a sentence. This means that tiles then move off-screen and the whole thing loses its illusion of being transported into the pipe. I think this can be fixed by adjusting the scrollbar position fully left when the lever is pulled, but I'm not certain that will work.

Another minor issue is that the distance each tile moves to the left is not exact. So after animating a ton of words, you'd start to see the tiles drift to the right ever so slightly because we are moving just shy of the distance from one tile to another.

#32