Resize/reposition various elements in the sentencebuilder scene
I was hoping to bring more attention to the sentence bar/conveyor belt rather than the word bank, seeing as that is the primary use of this scene. It could probably be emphasized more, but this should be a pretty noticeable shift in emphasis that I'd like to explore before committing to anything more major.

I've also increased the sizes of the tile dropzone areas for the sentencebar and wordholder to make it easier to drop tiles in them properly.
#30