Fix scrollbar not reaching the bottom of the wordbank
It turns out that the numOfRows calculation was a division of two integers that we were trying to round up, which was problematic because dividing two integers just shaves off any decimal points, so our Mathf stuff would appear to always round down. I've stored the words.Count as a float so that the division would briefly hang on to the decimal places so that Mathf.CeilToInt could properly round up for us.

I've also included a fix for the sentence bar vertical animation that fixes the tiles not going far enough to disappear. (as a comment in the hopes that it doesn't prompt merge conflicts)
#30, #29