Implement two different animation methods for saved sentences
revealSentenceWordByWord will just pop in the words from the saved sentence one by one without an animation. I plan on making the individual words slide in, but I'm not sure how to get them to position properly when they are all separate objects.

revealSentenceAnimation instead makes the whole sentence move as one big block, where each word is revealed approximately in time with the TTS voice.

I've also moved the MostRecentSavedSentence game object higher at the top of the canvas hierarchy so that the words in the sentence are hidden behind the pipe as they come in. (although a bit of the bottom of the text can be seen and should be fixed if this we end up liking the animation)