Implement scrollbar movement to match TTS and highlighting
Also:
Made the word tiles in the sentence bar slightly less wide.
Made the sentence bar slightly taller than the words it holds.
Made the sentence bar's slider properly shrink when dragging words out of the sentence bar.
Made the "reveal" animation at the bottom wait a bit before revealing to leave time for the pipes to animate in the future.
Made both the TTS button and submit sentence button shift the scrollbar left when they start.
Important note:
The functionality of the matching scrollbar movement to TTS is based on the assumption that exactly 6 tiles will fill up all of the sentence bar. If we end up changing around the sentence bar size or the sizes of the word tiles it houses, we'll have to make sure those changes are reflected in TextToSpeechHandler.cs