Implement re submission of completed sentences to sentence bar
The current implementation doesn't allow for positioning the resubmitted sentence in between words already in the sentence bar. It always appends to the end. If it turns out that we think it's important to be able to place the sentence in between word tiles, that can arranged.

This also fixes an unlisted issue where clicking on word tiles in the sentence bar and word holder wouldn't actually play the TTS. Turns out that when we'd added the voice modifiers, I'd forgotten to add the TTS handler component to them, and a script that thought it would have it was throwing a null reference exception. 

fixes #38