Adjust the number of tiles read before scrolling the sentence bar
Previously the sentence bar didn't begin scrolling until TTS reached the last word in view, where it would then reveal only the next word. This commit changes things so that we instead begin scrolling (if scrolling is possible) after reading 4 word tiles, (of the 6 in view) so that kids are afforded the context of the nearby words.

fixes #36 #37