Create word tile image for the submitted sentences
I've also made it so that the sentence tile is above everything else, with the exception of the sentence view. Because of some funky behavior with eventData.hovered, the dragged object needs to be below the sentence view in order to actually detect it. This is probably because the dragged object blocks raycasts otherwise, but I'm still looking into it.