Implement raycastTarget updates on drag method
This fixes the previous quirk of the system where eventData.hovered only worked when the object being dragged was underneath the sentence in the game view. The reason that was happening was because eventData.hovered was relying on raycasts to find things in the hovered stack that were being blocked by the dragged object.