Re-work leantween animation to fit new sentence set up
Now lean tween animates the whole collection of parent and child objects rather than just the text component.

Also added a white UI image behind the pipes to mask the sentence tile animating in. (so it looks like the tile is coming out of the pipe rather than off screen)