Change implementation of revealSentenceAnimation to hopefully be faster
I've stored the approximate time it takes for tts to speak through everything in a variable so it only needs to be calculated once. I'm wondering if maybe the weird stuttering lag was from having passed tts.getApproxSpeechTime to the lean tween moveX function. I could see it maybe having to invoke that method a bunch of times when any part of the moveX function that wanted to know how long the duration of the animation is supposed to be would have to run that method each time.