Only show sentences saved today
This change is a first exploration of filtering which sentences to show. By default, we will only show sentences from today. The other sentences are not deleted, but they will not show by default. We might wish to explore other defaults, but this seemed like a good place to start (and seemed realistic about what we might want).