Refactor JSON decoding
The code for reading in context packs from the JSON files was pretty challenging to read and follow. I switched it from using locations in the JSON to using keys from the JSON to access the various elements as much as possible. 

Currently, there are still some places that I think will need to be refactored further. For example, there is still an "offset" that is used to get the right spots in the JSON for the word packs. That's something I will probably change when I change the schema for the JSON. 

I'm concerned that the notion of contextPackId and wordPackId might be used oddly here in AddWord. There may be places we used an ID when the name would make more sense. Also, the way that words are added to the word list to only add one instance of a word seems to be problematic. Words that are already in the list will not be attributed to multiple contexts, which may impact filtering by context pack later. What we probably want is to have words only be displayed one time in the word bank, but to have (potentially) multiple entries so that they would show up in all appropriate filters.