Implement rough conveyor belt animation
Currently only pulls from two sprite images, so it looks a bit rough. Planning on adding one or two more to hopefully smooth that out.
Also planning on animating the pipes as the words get sucked into them, and changing the size of the word tiles as they are getting sucked in next.

Issue #44