Fixes the highlighting bug that kept tiles yellow after double clicks
The previous implementation was grabbing the image color as it was being asked to highlight, so if it was already in the process of highlighting it would grab the yellow color as its original and get stuck there. Original color determination now takes place on startup rather than when it's called.

Also cleaned up some old comments and removed the original ClearTiles() method that was causing weird lag spikes on mac. Some SentenceTile highlighting methods that weren't being used were also removed.

Side note: null reference errors on the storybuilder side of things are fixed in another branch.

Closes #47