Tweaks to pipe animation playback speed
Add a couple duplicate frames to start and end of animation. Trying to get the opening of the pipe to be more pronounced in the animation by having it last longer than the other frames.

Also gave the ProcessingTile condition to the PipeAnimation transition to PipeIdle. Previously it would play the animation and then switch back to idle (while also playing the pipe animation) regardless of whether it was told that tile processing was done, so the whole thing looked choppy and weird.

#44