Add transition animation between sentence and storybuilder views
This implements a change scene button that moves the camera view up and down depending on which view you're in using leantween. In this version, both canvases remain active to promote a feeling of cohesiveness, but there also exists commented code that activates/deactivates canvases as you toggle views if performance ends up being an issue.

There's still some weirdness to dragging word tiles in the sentence builder view and sentence tiles in the storybuilder view, which will be addressed later.

Targets issue #60