Fix weird dragging behavior across both scenes
Both canvases are rendered in screen space overlay mode when focused (focused meaning not changing scene) on, so that the existing drag functions don't have to account for camera position.

This fix introduces some minor weirdness of its own, in that when it changes scenes and toggles between screen space overlay and world space for the animation, there is a slight jitter as the different game objects positions are fit to the screen size.

Changes target issue #60