Begin implementation of wordcounting
This also includes creating a UserData class that consists of a (string) username and a dictionary<string, int> that will be used for database entries in the future. The only place we are calling the UpdateWordCount function is from clicking on the word tiles by themselves. Next on the agenda is adding word count tracking to sentence clicking and the storybuilder side of the world.

Targets issue #62