Import Json serialization package
This commit brings in Newtonsoft.Json-for-Unity (https://github.com/jilleJr/Newtonsoft.Json-for-Unity.git) so we can serialize and deserialize the dictionary used to store words and their counts. 

Currently properly tracks words and their word counts and outputs them to the console in simplistic Json formatting when a word tile is clicked on.

Targets issue #62