Initial test communication between WordRiver server and unity
This ports over the "User" and "Learner" datatypes from WordRiver, while also adding an initial test Get request that has Storybuilder grab a (hard-coded) User from the server and output their name to the debug console. 

A couple things to consider:
Right now it appears that the only way to get a user is by having their respective AuthID, and I'm not sure how we would get that info on the Unity side of things. It seems like we may want to create a GetUserByUsername() function on WordRiver's side or implement firebase login in Unity.

Targets issue #62