Implement Learner Login scene
This commit creates a scene after user login that allows people to choose which learner they are from all of the signed in user's list of learners. Currently the selected learner is not being passed to sentence builder. That will likely be the next step.

Also adds the Learner and User classes which both mirror their Wordriver counterparts of the same name. 

Targets issue #62

Co-authored-by: Kristen Lamberty <lamberty@morris.umn.edu>