Implements local, learner specific logging of wordcounts
This is only for the sentencebuilder portion of the scene for now.

Also adds the LearnerData class which contains all of the fields that will (probably) be relevant for communicating with the LearnerData collection once that gets made. (Learner object Id, learnerName, wordCounts)

WordCount updates are handled by calling the WordCountHandler script attached to the prefab gameobject of the same name.

Targets issue #62