Implement Get and Post requests for LearnerData in Unity
This commit finally has get and post requests working for learnerdata within the sentencebuilder scene. On the storybuilder side of things, this doesn't work yet.

In this implementation Unity only forms one get request upon learner selection, so if something is added to the learnerdata collection after logging in (outside of unity), Unity won't know about it.

The PostLearnerDataToServer() method still needs to be refactored to account for request errors like its Get counterpart.

Targets issue #62