Implement storybuilder scene wordcount tracking/server comms
This now allows us to track wordcounts throughout the entire storybuilder project! 

Note: The tracking isn't perfect in that it is incremented regardless of whether the entire sound plays out or not. So if someone rapid fire clicks on 3 different tiles and only actually listens to the final tile, they will be counted as having listened to all three anyway. Given the inherent uncertainty of learner engagement with the software at a more general level anyway (how can we know if they are really paying attention/listening when the sound plays?) this doesn't seem like a super huge issue.

Targets issue #62