Implement session length tracking in the LearnerData class
This commit creates logging for a dictionary of string, float where the string is the date/time upon entering the sentencebuilder scene and the float is the time the session lasted for in seconds.

Currently session time is only ever logged once, upon termination of the application. This doesn't account for IOS, because they usually suspend the application rather than killing it. (although there is an OnApplicationPause() method for that)

TODO: improve the formatting of the elapsed session time by converting the dictionary to a string, string format, and including an hour/min/sec breakdown of the time.

Targets issue #62