Implement local storage of learner sprites
This change was made to reduce the number of calls we have to make to the firebase server. Now instead of requesting each learner's icon every time we load the learner login scene, they are only requested the first time each session.

Also added a button to the learner login screen that will request learner icons even if they are already stored locally in case they need to be refreshed.

In implementing these changes, I also noticed that changing firebase users wasn't working properly previously. This was fixed by adding await to the firebaseAuth.SignInWithEmailAndPasswordAsync() so that we didn't load the learner login page before properly updating the firebase user.

Targets issue #62