Refactor local storage of context packs/etc and reimplement server calls
This swaps the storage location of LearnerData/ContextPacks/LeanerIcons to Application.persistentDataPath, which should work for IOS/Android/Mac/Windows as Unity changes the storage path based on the OS.

Also made a bunch of the buttons a lot bigger; using the app on a phone was quite painful with how small things were before. Context pack filters and letter filters for the wordbank should also be increased significantly.

Additionally, the end session times for LearnerData are now logged upon changing learners or logging out, as OnApplicationQuit() never seemed to fire from initial testing.