Re-implement/fix sentence storage in storybuilder scene
Previously sentences weren't being stored in the Application.persistentDataPath and were thus not being used in anything but PC OS's. This changes their storage location to that persistent path/Resources/Sentences. No communication with the server occurs for sentences; storage is entirely local.

This also fixes an issue where sentences weren't actually updating in the sentence bank when we swapped to the storybuilder canvas. This was because the sentence bank population logic only occurred upon initialization, which only occurs once now that we've combined the storybuilder/sentencebuilder scenes into one. We now rely on OnEnable/OnDisable, where sentencebuilder => storybuilder triggers OnEnable and vice versa. Current implementation repopulates this sentence bank each time we re-load it rather than ensuring duplicate sentences aren't added, which may be something to consider re-factoring in the near future.