Add error message to user login screen for failed logins
The UserLogin error message implementation is a bit weird in that it relies on updating a boolean loginFailed variable upon failing to login that Unity's Update() method then checks for each frame. This is because the login method takes place outside of the main thread, where the Unity API is not available. This could probably be made more efficient by checking less frequently in the future.

This also removes some old sentences we'd had stored that aren't accessible in android builds of the project.