Fix saved sentences in storybuilder not preserving word form changes
SavedSentence's were refactored to include a string of the sentence text and a list of the word strings that make up said sentence. This was to get around the issue of being unable to change word.word without also changing all other word tiles with the same script attached.

Also changed the serialization for sentences from binary to regular old json files.