Initial implementation of sentence server request POST/GET methods
This allows us to track a learner's submitted sentences, the time they were submitted, and the userId of the user that they belong to. 

A weird quirk of this implementation is that the actual server call that gets a learner's sentences from the database on startup is attached to the ChangeScene.cs script, where it would make more sense to have it attached to the SavedSentenceBank.cs script. This is because Unity's Start() method won't fire on SavedSentenceBank.cs until we've changed scenes (it starts disabled). This is a problem because changing scenes prompts loading sentences from local files which could then cause conflicting read/write operations or even weird race condition bugs.

TODO: add used context pack information to the SavedSentence class. This would help track which context packs a learner most interacted with.