Refactor sentence submission to track context packs used
This entirely removes the ServerContextPack class and instead combines its fields with the older ContextPack class. This combination originally seemed like it would be a problem because of serialization issues/not wanting to track certain fields, but simply adding the [NonSerialized] tag to fields allows the class to deviate a bit from its server equivalent.

In this implementation the Word class now knows what contextPackId it's from in unity and the web client, without storing that information in the database. Instead, words simply hold their context pack Id for sentences to be able to figure out what context packs were used in their creation.