Import necessary files for IOS build to (mostly) function
There's still some weirdness in that the voice select buttons don't seem to be working in two ways. Firstly, the buttons don't seem to spawn in until you've logged into the user twice, and even when they do properly spawn in, while they all have different names, they all appear to be the same voice.

NOTE: Wordholder is known to be broken; there is a branch in the pipeline that will fix this.