Fix context packs not loading properly
Previously we were getting null reference errors when checking the "word" field of our JSON objects because of a serialization issue between wordriver's "word" and storybuilder's "baseword". For now I've just added a JsonProperty tag that will serialize storybuilder's baseword into wordriver's conventions, but it may be a good idea to also refactor on that side of things in the future.