Fix bug sentence tile duplication bug
Previously when dragging sentence tiles back into empty space in their bank the tiles would "stick" there instead of deleting like they were supposed to. This was because the dropzone script was attached to the sentencebanknew game object, which scales with the number of sentence tiles it has. So when there were only a few sentences, there was empty space beneath it that didn't trigger any of the switch cases. To fix this, the script has been added to the sentencebankscrollview game object instead, as its size is static regardless of the number of sentences.