Add display for current logged in learner
This adds the learner's profile picture and name to the bottom left of whatever scene the user is in. Slightly re positioned the open menu button so that it remains fully on screen when switch back and forth between scene. 

Also removes a silly optimization with changing around Time.fixedDeltaTime and moved the functions that grab learner sprites to the LearnerIconStorageHandler class.