Implement context pack icon support
This is more-so a first iteration of this feature. It currently calls the server everytime (no local storage) and the actual passing around of data from firebase to the sprite of the game object is clunky.

It does however seem to properly place the context pack icon onto the pack filter button.

Also added constructors for LearnerData to clean up some of the assignment clutter in ServerRequestHandler.cs