Swap RTVoice prefab to UserLogin scene
Previously I'd thought there were issues with not enabling standalone TTS breaking RTVoice stuff but it turns out that flipping that setting on and off was just placing the RTVoice prefab in the UserLogin screen. Having it there seems to make things work properly so I've moved it there manually now.