Implement initial story saving to database
This saves the stories users create as a list of pages that contain a page number (int) and a list of sentences (string) that were on that page. This implementation currently allows the submission of empty stories (where no sentences or even pages exist), but will likely be given checks on whether the stories actually contain anything later. These stories are currently unable to be titled.