Work on adding new word forms menu
This commit begins the process of adding a new forms menu. This commit comments out the code for the old word forms menu and adds code for the new one. Currently, the new word holder just displays the base word, and does not display word forms or have any other real functionality.