Get word forms to destroy on deactivate
This commit makes so that the word form game objects destroy themselves when the word forms pop up window is closed. Right now, the word forms display in the correct alignment, but if there are too many then they run off the window. Additionally, if the user does not close the window before dragging a new word in there, then the word forms from the previous word may still appear in the window.