Prevent user input while sentence is being built
This commit adds a transparent panel that activates while a sentence is being built. This panel blocks user input so the user cannot touch/click on anything while the sentence is being built. This was done because the user could add words while the sentence was being built. The user could also click the text to speech button which would override the tts and result in the  whole sentence building becoming off.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>