Update Pitch Slider Sprites
This commit updates the voice pitch images so that the increase pitch image is a chipmunk, and the decrease pitch image is a cow. It also switches the voice sliders in the story builder canvas to match the ones in the sentence builder canvas. It adds the images to both of the sliders in the story builder canvas as well.

Co-authored-by: Jeklund99 <eklun124@umn.edu>