Move story name prompt and add button to click to submit
This commit moves the story name prompt up so that it looks better on an IPad. Additionally, a button was added that the user can click in order to submit their chosen name. This was done due to Unity not detecting the enter key being pressed on the onscreen keyboard on the ipad.