Add option to stop tts when speaking story
This commit adds a stop sign that replaces the speaker button after the speaker has been clicked so that the user is able stop the sentence from being read completely. It does stop the learner data from being updated incorrectly (counting all the words in the story even if they have not been read).

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>