Refine getSentences query
This commit refines the getSentences query in the ServerRequestHandler script so that we only get the sentences from the most recent session rather than all of the sentences. Also changes the Story.cs script to include a time submitted field. This was done to make this possible on the word river side.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>