Fix issue where user would get sentences from previous story
This commit fixes the issue where we would get sentences from the most recent story. This required changing the sentence/story controllers on the Word River side of the project. This commit also updates the saved sentence structure so that deleted sentences are updated in the server properly. Now the deleted sentences no longer reappear when unity is restarted.

Co-authored-by: Natasha-Zebrev <zebre001@morris.umn.edu>