Rename TextToSpeachButton, prevent input while sentence is being read
This commit renames the "TextToSpeachButton" and associated script to "SpeakSentenceButton" to more accurately reflect what the button is doing. This commit also replaces the "isSpeakingSentence" with "isSpeaking" as having both of them was redundant. This commit also fixes an issue where the user would be able to interact with several game objects while the TTS was speaking a sentence that was in the conveyor belt.