Disable automatic word form popup when dragging tile into word holder
This commit comments out the line of code that makes it so that when you drop a tile into the word holder, the word form selection screen pops up. This forces the user to push the button on the word form holder in order to access the word forms. Leaving the commented out line in the code for now in case we wish to restore this functionality.