Fix filtering buttons
Last commit broke the filtering buttons. This was because the filtering buttons use the "find" method to acquire a reference to the FilterController script. This script was attached to the game object that functions as the hide filter buttons. So, I re-activated that game object but just disabled the button and image components, so only the script components of that game object are enabled.