Work on context pack icons
This commit gets context pack icons working (for the most part). Issue is with the AlreadyHaveAppropriateContextPackIcons method in the ContextPackHandler script. That method doesn't always run for each pack. It just returns true if it finds one icon that is already in storage. Need to change it so that it runs for every pack.