Get context pack icons to work
This commit gets the context pack icons to load every time. However, there appears to be a bug that occurs where the context  packs will load twice, so there will be two buttons for each context pack. Not quite sure yet what causes this, but it is fixed by updating the word bank.