Update IP Address
Somewhere along the way, an IP Address was committed and pushed to GitHub. This removes the IP, replacing it with localhost. Using localhost is most common for development purposes, but in production it is necessary to create a build of Story Builder that has the actual IP address that is hosting the Javalin server for Word River.

Co-authored-by: El Anderson <and10393@umn.edu>